requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
python-multipart
jinja2
pymongo
uvicorn[standard]